% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_data_validation.R
\name{check_counts_nonneg_integer}
\alias{check_counts_nonneg_integer}
\title{Check that counts are non-negative integers}
\usage{
check_counts_nonneg_integer(catch_data)
}
\arguments{
\item{catch_data}{See \code{\link[=validate_cpue_input]{validate_cpue_input()}}.}
}
\value{
Character vector of failure messages.
}
\description{
Check that counts are non-negative integers
}
\keyword{internal}
