% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_data_validation.R
\name{check_reach_id_consistency}
\alias{check_reach_id_consistency}
\title{Check that every reach_id in catch_data exists in reach_metadata}
\usage{
check_reach_id_consistency(catch_data, reach_metadata)
}
\arguments{
\item{catch_data}{See \code{\link[=validate_cpue_input]{validate_cpue_input()}}.}

\item{reach_metadata}{See \code{\link[=validate_cpue_input]{validate_cpue_input()}}.}
}
\value{
Character vector of failure messages.
}
\description{
Check that every reach_id in catch_data exists in reach_metadata
}
\keyword{internal}
