% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_population_estimation.R
\name{.zippin_no_var}
\alias{.zippin_no_var}
\title{Large-sample variance of the Zippin abundance estimate}
\usage{
.zippin_no_var(N0, p, k)
}
\description{
Large-sample variance of the Zippin abundance estimate
}
\keyword{internal}
